file handlers limit
