+++ src/core/util/fs_util.cpp
@@ -22,6 +22,28 @@
 
 #include <sys/param.h>
 
+bool turi::fs_util::upgrade_file_handle_limit(size_t limit) {
+#ifndef _WIN32
+  struct rlimit rlim;
+  rlim.rlim_cur = limit;
+  rlim.rlim_max = limit;
+  return setrlimit(RLIMIT_NOFILE, &rlim) == 0;
+#else
+  return true;
+#endif
+}
+
+int turi::fs_util::get_file_handle_limit() {
+#ifndef _WIN32
+  struct rlimit rlim;
+  int ret = getrlimit(RLIMIT_NOFILE, &rlim);
+  if (ret != 0) return 0;
+  return int(rlim.rlim_cur);
+#else
+  return 4096;
+#endif
+}
+
 bool is_hidden(const std::string path) {
   if (path.length() && path[0] == '.') {
     return true;
